Use bower to manage vendor assets
Add new directory /bower_components for managing third party
stylesheets and javascripts. This dir is now added to the sprockets
asset path in config.rb. This commit also upgrades to Foundation 5
and drops the zurb-foundation gem in favor of the bower install.
The highlightjs dependency is also moved to bower.